I think I finally have a single app state working.
The state is stored in the IYSStateService and bound to the template via
IYSStateController.

Methods have been added to the state service for updating the active
items.

Just need to watch for changes in the carousels and call the appropriate
update function to get the states to match and make magic happen.